fixed fetch typo, moved scripts to head, fixed destroy player in backend

+++ drawerv2-backend/app/controllers/players_controller.rb
@@ -23,9 +23,12 @@ class PlayersController < ApplicationController
 
     def destroy
         player = Player.all.find_by(id: params[:id])
-        player.destroy
-        render json: {message: "Mess with the best, die like the rest"}
-        render json: player
+        if player
+            player.destroy
+            render json: {message: "Mess with the best, die like the rest"}
+        else
+            render json: {message: "Delete Failed"}
+        end
     end
 
     private
